restorer: extract hardlinks index from restic package
